rename: view폴더 생성후 InputView, Output를 view폴더로 이동
